requirements.txt
numpy==1.20.3
torch==1.8.1
matplotlib==3.4.2
omegaconf==2.0.6
pandas==1.2.4
pytorch_lightning==1.3.4
hydra_core==1.0.6
tqdm==4.61.0
torchaudio==0.6.0
librosa==0.8.1
pytest==6.2.4
seaborn==0.11.1
